Add placeholder NPC, knife
Throw the knife with space. It only goes right for now, but that will be
changed. It destroys itself when it hits a wall, and destroys both
itself and the NPC when it hits an NPC. Going outside the boundaries of
the world does nothing to the knife, and you won't be able to throw
another one.