Make code more object-oriented
To the degree that javascript can manage that. The player and HUD-text
are now function objects, the player taking their knife along for the
ride. Also some miscellaneous code changes, most notably to input setup.

Bug fixed: When knifing something (walls, players) that is not the first
in a group, the game no longer crashes.