Update throwKnife to be player-dependent
Now it works with more than just the client player, which was the entire
point of putting that function in the global scope to begin with. Also
assigned masses to objects (based off of the Unity project), but that's
wonky as hell. I'm thinking we'll need some custom overlap functions for
normal wind behavior.